# Conflicts:
#	biz-module-flink-pro/src/main/java/com/yqwl/datamiddle/realtime/common/MysqlConfig.java

+++ biz-module-flink-pro/src/main/java/com/yqwl/datamiddle/realtime/bean/DwmSptb02.java
@@ -595,7 +595,7 @@ public class DwmSptb02 implements Serializable {
      * 返单大序号  number 转 int
      */
 
-    private Integer NFDDXH;
+    private Long NFDDXH;
 
     /**
      * 发车站台
